+++ llvm/lib/Target/NVPTX/NVPTXTargetTransformInfo.cpp
@@ -280,7 +280,7 @@ static Instruction *convertNvvmIntrinsicToLlvm(InstCombiner &IC,
       //
       //   - nvvm_cos_approx_{f,ftz_f}
       //   - nvvm_ex2_approx(_ftz)
-      //   - nvvm_lg2_approx_{d,f,ftz_f}
+      //   - nvvm_lg2_approx_{f,ftz_f}
       //   - nvvm_sin_approx_{f,ftz_f}
       //   - nvvm_sqrt_approx_{f,ftz_f}
       //   - nvvm_rsqrt_approx_{d,f,ftz_f}